refactor: run WAF solver entirely in a bundled worker
Move all WAF network I/O and CPU-bound work off the main event loop and
fix a production-only breakage in how the solver worker was spawned.

- Run the full flow (page fetch, challenge discovery, deobfuscation,
  fingerprint, proof-of-work, verify) inside one Bun Worker
  (solver-worker.ts) so brute-force PoW never blocks the main thread.
- The 202 challenge body is now parsed in place instead of re-fetching
  the URL, eliminating a redundant request to the bot-protected endpoint.
- solver.ts becomes a thin main-thread client owning the token/config
  cache; it passes cached state into the worker and folds the result back.
- Bundle the worker via standaloneBundles() to
  .output/server/workers/waf-solver-worker.js and load it with the same
  .js (prod) / .ts (dev) switch the other workers use. The previous
  extract-worker was spawned from a never-bundled .ts path and would fail
  in the Nitro build. Verified the file lands in the Docker image and the
  runtime path resolves to it.
- Split deobfuscation into a pure deobfuscate.ts (import.meta.main CLI for
  offline testing); add messages.ts (worker contract) and pageMarker.ts
  (shared __NEXT_DATA__ marker). Update the README.
